Prevent host's node_modules from being mounted
In development mode, we mount the host's backend source in order
to watch for changes and rebuild as needed. Previously, the host's
node_modules directory would overwrite the version built in the
container. If the host's dependencies were out-of-sync with those
in package.json, it could cause problems when running the container.
To avoid potential problems, we now ensure the container's
node_modules is always used.