Add Typescript Support/Compilation (#305)
* Add @TanStack Tablev8 to the NPM package

Note that this gave a bunch of warnings due to TanStack requiring
more recent versions of various core React modules. Unsure if this
will cause unexpected issues later.

* Update Bibliography Table to TanStackv8

Proof of concept replacement of most functions with the newer
replacements. Does not further change any site structure, and
notably does not fix/update the export functionality.

* Add @tanstack/react-table to the frontend node-modules

turns out there's separate NPMs. Also, this one didn't have
errors for outdate dependencies.

* Fix compilation errors

The table now displays, but there's still a few things going weird
with it. I think the column definitions aren't matching and are
refusing to render their contents because of it.

* Implement proof of concept v8 Table for Biblio

Version is not ready for merging into final codebase.
Currently lacks filtering and exporting, due to needing
to implement custom components under the new system.

* Remove npm install from base directory

I could have reverted the specific commit but this seemed
less likely to cause a random error.
Not a feature, just fixing a mistake from before.

* Add Typescript and associated dependencies.

Also automatically updated some other package items in the process.
I had to select certain versions of many of these components because
later versions require Node v14 or later.
Configuration of the typescript compiler file appears to work...

* Create functioning tsx filter component for table

Integrated as a proof of concept into the tanstackv8 table
conversion that I was working with before.
Currently not serving as a proper fuzzy filter and some features
could be otherwise improved.

* Update BibliographyTable to display old version

Also tried to fix some stuff about the new one, but
it's not quite working yet so we will delay any changes.